*: 4.14.14 guards against GCPMintModeRoleAdmin
[1] contains [2].  Raising minor_min ensures the next 4.15 release
built from that advice will not include update recommendations from
any unguarded 4.14, and we can declare that 4.15.z as the fixedIn
version and stop carrying the risk.  But rc.8 was built before we had
the updated advice in place, and still includes the older 4.14.z as
update sources:

  $ oc adm release info -o json quay.io/openshift-release-dev/ocp-release:4.15.0-rc.8-x86_64 | jq -c .metadata.previous
  ["4.14.10","4.14.11","4.14.12","4.14.13","4.14.14","4.14.9","4.15.0-ec.0",...

So extend the risk declaration to rc.8 as well.

[1]: https://amd64.ocp.releases.ci.openshift.org/releasestream/4-stable/release/4.14.14
[2]: https://issues.redhat.com/browse/OCPBUGS-28231