blocked-edges/4.15.0-EarlyAPICertRotation: Fixed in 4.15.2
[1] contains [2], which addresses the CA-rotation trigger for clusters
born in 4.7 or earlier.  We are still working on getting 4.15.z
systems to respond gracefully when the rotation happens,
e.g. backporting the changes in [3] to 4.15.z, but as far as we can
tell, 4.15.2 clusters will no longer try and roll the api-int
Certificate Authority, so we're no worse off there than in the 4.14
releases users would be updating from.

[1]: https://amd64.ocp.releases.ci.openshift.org/releasestream/4-stable/release/4.15.2
[2]: https://issues.redhat.com/browse/OCPBUGS-30304
[3]: https://issues.redhat.com//browse/OCPBUGS-25821