blocked-edges/4.14.*-OVNInterConnectTransitionIPsec: Declare risk
The PromQL is:

* If we know IPsec is enabled, we're exposed, or...
* If we know IPsec is not enabled, we're not exposed, or...
* If are OVN, but aren't sure if IPsec is enabled
  (e.g. ovnkube_master_ipsec_enabled scraping is failing), we might be
  exposed.  -1 will cause an evaluation failure [1].  Or...
* If we know we are not OVN, we are not exposed, or...
* If we aren't sure we're on OVN (e.g. apiserver_storage_objects
  scraping is failing), we might be exposed.  Returning no time series
  will cause an evaluation failure.

The label_replace makes it easier to tell when we're in the "we know
IPsec is not enabled" case.

The max_over_time avoids hiccuping if metrics are interrupted; see
5e4480fa76 (blocked-edges/4.14.*-AzureRegistryImagePreservation: Look
for active registry use, 2024-02-09, #4763).

I'm also adding _id="" to the queries as a pattern to support
HyperShift and other systems that could query the cluster's data out
of a PromQL engine that stored data for multiple clusters.  More
context in 5cb2e93728 (blocked-edges/4.11.*-KeepalivedMulticastSkew:
Explicit _id="", 2023-05-09, #3591).

Generated by creating the 4.14.0 file by hand and copying it out to
the other 4.14 releases with:

  $ curl -s 'https://api.openshift.com/api/upgrades_info/graph?channel=candidate-4.14&arch=amd64' | jq -r '.nodes[] | .version' | grep '^4[.]14[.]' | grep -v '^4[.]14[.]0$' | while read VERSION; do sed "s/4.14.0/${VERSION}/" blocked-edges/4.14.0-OVNInterConnectTransitionIPsec.yaml > "blocked-edges/${VERSION}-OVNInterConnectTransitionIPsec.yaml"; done

[1]: https://github.com/openshift/enhancements/blob/4668a0825c59739dfafd2ae661c16cf30f540946/enhancements/update/targeted-update-edge-blocking.md?plain=1#L119