blocked-edges/4.16.*-FlowControlV1Beta3Migration: Declare new risk
034fa0182a (blocked-edges/4.12.*: Declare AWSOldBootImages,
2022-12-14, #2909) explains why we need to look for 4.9-or-earlier
instead of looking for the 4.8-or-earlier condition this risk is
associated with.

I'm also adding _id="" to the queries as a pattern to support
HyperShift and other systems that could query the cluster's data out
of a PromQL engine that stored data for multiple clusters.  More
context in 5cb2e93728 (blocked-edges/4.11.*-KeepalivedMulticastSkew:
Explicit _id="", 2023-05-09, #3591).

Fixed in rc.4, because it has the new minor_min from f8316da482
(build-suggestions/4.16: Set minor_min to 4.15.17, 2024-06-06, #5352):

  $ oc adm release info quay.io/openshift-release-dev/ocp-release:4.16.0-rc.3-x86_64 | grep Upgrades
    Upgrades: 4.15.11, 4.15.12, 4.15.13, 4.15.14, 4.15.15, 4.15.16, 4.16.0-ec.1, 4.16.0-ec.2, 4.16.0-ec.3, 4.16.0-ec.4, 4.16.0-ec.5, 4.16.0-ec.6, 4.16.0-rc.0, 4.16.0-rc.1, 4.16.0-rc.2
  $ oc adm release info quay.io/openshift-release-dev/ocp-release:4.16.0-rc.4-x86_64 | grep Upgrades
    Upgrades: 4.15.17, 4.16.0-ec.1, 4.16.0-ec.2, 4.16.0-ec.3, 4.16.0-ec.4, 4.16.0-ec.5, 4.16.0-ec.6, 4.16.0-rc.0, 4.16.0-rc.1, 4.16.0-rc.2, 4.16.0-rc.3

and the fix for [1] is in [2].

[1]: https://issues.redhat.com/browse/OCPBUGS-34492
[2]: https://amd64.ocp.releases.ci.openshift.org/releasestream/4-stable/release/4.15.16