blocked-edges: Declare OpenStackAvailabilityZoneOutOfRange
Regressions landed in 4.16.0-ec.6 [1,2], 4.15.15 [3] and 4.14.29 [3].
Generated by creating the 4.14.29 risk and copying it around with:

  $ for V in 4.14.30 4.15.{15,16,17,18} 4.16.0-{ec.6,rc.{0,1,2,3,4}}; do sed 's/4.14.29/${V}/' blocked-edges/4.14.29-OpenStackAvailabilityZoneOutOfRange.yaml > "blocked-edges/${V}-OpenStackAvailabilityZoneOutOfRange.yaml"; done

I've gone with 'from: .*', because I expect a fix to move back
quickly, but we can revise to allow updates within affected releases
(e.g. 4.15.15 -> 5.15.16) in follow up work, if we want those flowing
while we wait for the fix.

[1]: https://amd64.ocp.releases.ci.openshift.org/releasestream/4-dev-preview/release/4.16.0-ec.6
[2]: https://issues.redhat.com/browse/OCPBUGS-30951
[3]: https://issues.redhat.com/browse/OSASINFRA-3500