README: Explain scoping vs. OCP policy and support
To reduce the chances that an OpenShift-cluster administrator reading
the README jumps from "this API allows the graph-data admins to do
things like dropping update edges" (which is true) to "so maybe
tomorrow my 4.15.0 to 4.15.2 recommendation will be blocked in
stable-4.15!" (which is not true).  That won't happen because [1]
says:

  Beginning in OpenShift Container Platform 4.10, when update risks
  are confirmed, they are declared as Conditional Update risks for the
  relevant updates.

which is a policy commitment constraining update service changes.
Hopefully the new wording adds some clarity on the space between "how
the tools could be used" and "how Red Hat will use the tools".

[1]: https://docs.openshift.com/container-platform/4.15/updating/understanding_updates/understanding-update-channels-release.html#conditional-updates-overview_understanding-update-channels-releases