blocked-edges/4.16.*-ServiceAccountContentionSecretLeak: Declare risk
The regression came in via [1], landing in rc.3 [2].  I wrote the
4.16.0 risk by hand, and copied it out to the other 4.16 (by excluding
Engineering Candidates and the first few Release Candidates) with:

  $ curl -s 'https://api.openshift.com/api/upgrades_info/graph?channel=candidate-4.16&arch=amd64' | jq -r '.nodes[] | .version' | grep '^4[.]16[.]' | grep -v '^4[.]16[.]0$\|-ec[.]\|-rc[.][0-2]' | while read VERSION; do sed "s/4.16.0/${VERSION}/" blocked-edges/4.16.0-ServiceAccountContentionSecretLeak.yaml > "blocked-edges/${VERSION}-ServiceAccountContentionSecretLeak.yaml"; done

0cc6e3c475 (./hack/release-ga.sh 4.16, 2024-06-27, #5490)'s PreRelease
risks mean we don't need to bother with "but what about clusters on
rc.0 updating into 4.16.0?" unless we want to protect clusters that
are already (and still) running those very-prerelease versions.  I'm
not worrying about them for now.

[1]: https://issues.redhat.com/browse/OCPBUGS-33600
[2]: https://amd64.ocp.releases.ci.openshift.org/releasestream/4-stable/release/4.16.0-rc.3