blocked-edges/4.13.46-AzureSystemDLooping: Declare Azure-scoped risk
There is already an ARO-scoped AROSystemDLooping risk which has been
evolving:

* 2183f0c284 (Block edges for ARO clusters with systemd dependency
  loops, 2024-08-08, #5683).
* 2ef3ccd02f (Update AROSystemDLooping promql to have no errors,
  2024-08-09, #5693).
* 25b37ba861 (Update AROSystemDLooping risk based off testing,
  2024-08-14, #5721).

But we hadn't seen that impact in 4.12.recent to 4.13.recent non-ARO
Azure CI.  As we continued to test, the 4.12.54 [1] change seems to be
what's protecting clusters installed with 4.12.recent.  And [2] is
taking a fix back to 4.13.z that will allow the machine-config daemon
to update units without spooking systemd with dependency cycles while
the units are being updated.

[1]: https://issues.redhat.com/browse/OCPBUGS-30823
[2]: https://issues.redhat.com/browse/OCPBUGS-38295