hack/stabilization-changes: Add support for autoExtend
I'm not actually automatically extending risks yet, but that can come
later.  For now, I'm just giving the graph-data administrator a clear
message requesting the extension.

I've retroactively tucked this into version 1.1.0, because like
fixedIn, it's between graph-data admins and the stabilization script,
and not something that Cincinnati or later components are expected to
care about.

For ARODNSWrongBootSequence, Lisa told us on 2024-09-26 that we could
use that ARO ticket to delegate managing that particular risk to her.