WIP: blocked-edges/4.14.*-OVNTransitSwitchSubset: Declare 4.13-to-4.14 risk
The risk is briefly covered in the linked release note since
2024-09-18 [1].

I'm setting autoExtend, because unless some future 4.13.z addresses
this (unlikely, with 4.13.z going end-of-life on 2014-11-27 [2]), or a
future 4.14.z addresses this, we'll likely have to continue to carry
it forward to every 4.13-to-4.14 update we ever support.  This will
make life more difficult for customers who traditionally wait until
there's an update we recommend unconditionally.  All OVN cluster
administrators on 4.13 will need to read through the risk message and
linked docs, and decide for themselves whether they are comfortable
accepting the risk and updating to 4.14.

Generated by writing the 4.14.0 risk by hand, and then copying out to
all existing 4.14.z update targets with:

  $ curl -s 'https://api.openshift.com/api/upgrades_info/graph?channel=candidate-4.14&arch=amd64' | jq -r '.nodes[] | .version' | grep '^4[.]14[.][0-9]*$' | grep -v 4.14.0 | while read VERSION; do sed "s/4.14.0/${VERSION}/" blocked-edges/4.14.0-OVNTransitSwitchSubset.yaml > "blocked-edges/${VERSION}-OVNTransitSwitchSubset.yaml"; done

[1]: https://github.com/openshift/openshift-docs/pull/81666#event-14307870553
[2]: https://access.redhat.com/support/policy/updates/openshift#dates