channels/stable-4.16: Add 4.14.z
Folks on 4.(y-2).z may wish to go to 4.y but not qualify for the
Extended Update Support [1] or Extended Update Support Term 2 [2] life
cycle phases.  By adding 4.(y-2).z to stable-4.y, we can pivot
control-plane-only docs [3] to use stable-4.16 and similar in their
examples, to make it more clear that that paused-MachineConfigPool
update process is orthogonal to the EUS-ness of the life cycle phase
or channel name.  I'd initially done this for all the
currently-supported stable-4.y with:

  $ for C in channels/eus-4.1{2,4,6}.yaml; do sed 's/eus-/stable-/' "${C}" > "${C/eus/stable}"; done

But Subin suggested we start with just 4.16, and leave stable-4.14 and
stable-4.12 alone, so I've done that with:

  $ sed 's/eus-/stable-/' channels/eus-4.16.yaml > channels/stable-4.16.yaml

[1]: https://access.redhat.com/support/policy/updates/openshift#eus
[2]: https://access.redhat.com/support/policy/updates/openshift#eust2
[3]: https://docs.openshift.com/container-platform/4.16/updating/updating_a_cluster/control-plane-only-update.html