`OVNLocalnetWithNoSubnets`: ->4.17.[0-6] with virtualization
Address SDN-5485 / OCPBUGS-43454 which affects 4.17 until 4.17.7 where
it is fixed. We have not managed to identify a precise technical PromQL
to identify clusters with the exact problematic networking configuration
so I am proposing to use a coarse approximation of what clusters are
*likely* (based on the information available) to have such config (OpenShift
Virtualization installations identified by the hyperconverged-cluster-operator
CSV presence, which at the same time have at least one `network-attachment-definitions.k8s.cni.cncf.io`
resource instance.

Given the impact of the bug when hit (failure to create new workloads),
it is better to steer the group of potentially affected clusters towards
already available 4.17.7, even if we are not able to produce a sufficiently
targeted PromQL for the actual network configuration.