OCPBUGS-56251: Set 4.18.19 as the minor_min for 4.19 upgrade
Using Alertmanager apiVersion v1 in openshift-monitoring/cluster-monitoring-config or
openshift-user-workload-monitoring/user-workload-monitoring-config will cause
(InvalidConfigXXX) as OCP 4.19 is using Prometheus V3 and
Alertmanager apiVersions v1 is not supported in it.
With the change included in 4.18.19, it will set CMO upgradable=false if apiVersion: v1 is detected in the
config maps. Hence making it as minor_min version before upgrading to 4.19

Signed-off-by: Jayapriya Pai <janantha@redhat.com>