blocked-edges/4.18.*-CrunConflictsWithNVIDIA: Declare risk
Still some uncertainty in the impact statement, and it would be nice
to have NVIDIA and/or crun experts suggesting wording instead of my
guessing.  But it's fairly clear that something is happening on these
two releases, so get an initial warning out, and we can work on
wordsmithing later.