CONSOLE-4762: `ConsoleCrashOnMissingPlugin`: Declare on 4.19.(7-13)
If a Console Operator configuration resource references a plugin name
for which there is no corresponding ConsolePlugin resource, the Console
may crashloop after the cluster is updated to an affected version.

There seems to be no suitable metric to identify the issue, so we make
this an `Always` risk and will issue it once the version with a fix is
released.