graph-data.rs/src/check_signatures.rs: Limit concurrency when fetching signatures
By increasing the logging verbosity, the issue causing the e2e to fail
can be found.

A modified excerpt from the e2e job of a testing draft PR that increases
the logging verbosity of this section of the code:

- Failed to fetch signature from https://mirror.openshift.com/pub/openshift-v4/signatures/openshift/release/sha256=e1ed6d1672bf041f5b35c4cc6d0222bbf8126e070ec1d957753cc61ddf48082e/signature-1
  - Caused by:
    - 0: error sending request for url (https://mirror.openshift.com/pub/openshift-v4/signatures/openshift/release/sha256=e1ed6d1672bf041f5b35c4cc6d0222bbf8126e070ec1d957753cc61ddf48082e/signature-1)
    - 1: client error (Connect)
    - 2: tcp open error
    - 3: Too many open files (os error 24)",
- Failed to fetch signature from https://mirror.openshift.com/pub/openshift-v4/signatures/openshift/release/sha256=e1ed6d1672bf041f5b35c4cc6d0222bbf8126e070ec1d957753cc61ddf48082e/signature-2
  - Caused by:
    - 0: error sending request for url (https://mirror.openshift.com/pub/openshift-v4/signatures/openshift/release/sha256=e1ed6d1672bf041f5b35c4cc6d0222bbf8126e070ec1d957753cc61ddf48082e/signature-2)
    - 1: client error (Connect)
    - 2: dns error
    - 3: Device or resource busy (os error 16)"

Something has changed regarding the running container executing the
logic. The timeline corresponds to a commit that changed the cluster for
the cincinnati-graph-data presubmits configuration file to build02 from
build06. However, this may also be caused by some changes in how the
test containers are deployed. The end result is that e2e job is
overwhelming its perceived environment with its requests.

---

The commit introduces a limit to concurrently verified signatures.

---

The average duration of the last 70 e2e runs was about 25 minutes.

Trying different values once:

SIGNATURE_CHECK_CONCURRENCY	Result	Duration	Commit
1	✔️ Passed	54m3s	df3506a (different PR)
100	✔️ Passed	27m47s	ac05b61
300	✔️ Passed	30m17s	2dfb8d2 (reverted)

Thus, I am proposing to use 100 to leave some breathing space for
fluctuations, and since the value seems to be enough to finish
in a reasonable time.