blocked-edges/*-OSUpdateFailureDueToImagePullPolicy: Make message version agnostic
The risk was extended for some update paths to 4.20.0.

The currently existing sentence in the update risk message
"Please update to 4.19.16 or later where this issue is resolved."
will get displayed even for update path from 4.19.16 and 4.19.17 to
4.20.0, which does not make sense and may confuse cluster
administrators.

Make the message version agnostic by removing the information regarding
fixed versions.

An alternative is to create a new update risk with a different message,
as a specific update risk must contain the same message across
all paths.