hack/release-end-of-maintenance: Add a script to automate end-of-maintenance
Standardizing the feeder changes that remove a 4.y from stable-*.  An
example for 4.even (where end-of-maintenance is the beginning of
Extended Update Support) is d6157cc406 (channels/stable-*: Remove
4.14.z feeders from stable channels, 2025-05-08, #7227).  An example
for 4.odd (where end-of-maintenance is also end-of-life) is b2556281ee
(channels/stable-*: Remove 4.15.z feeders from stable channels,
2025-10-22, #8183).

The 'sed' expression for the stable-${MAJOR}.$((MINOR + 1)) and +2
channels is a straightforward filter replacement, as we have in other
release-* scripts.

The 'sed' expression for stable-${MAJOR}.${MINOR} is easy to explain:
it removes the entire 'feeder' block.  But the expression to deliver
that is tricky in the YAML-unaware 'sed':

* '/feeder:/,/^[^ ]/' selects an address-range starting with 'feeder:'
  and ending with the next line that is not indented with spaces
  (e.g. 'name:').
  * Operating on that address-range, '{//!d}' deletes all the internal
    lines in that address-range, but retains the bounding lines
    (because we want to keep 'name:', or whatever opens the subsequent
    YAML block).
* ';' wraps up the initial command and opens a new one.
* '/feeder:/d' deletes the 'feeder:' line we didn't delete in the
  first command.