blocked-edges/*-RuncShareProcessNamespace: hostNetwork:true is not required
No public evidence to link for this yet, but I'm hearing internally
that some shareProcessNamespace:true Pods that do not set
hostNetwork:true are also exposed.  I'm adjusting the message to
reflect that broader scope with:

  $ sed -i 's/in Pods where both shareProcessNamespace and hostNetwork are/in some Pods where shareProcessNamespace is/' blocked-edges/*RuncShareProcessNamespace*

Adding "some" also holds space for shareProcessNamespace:true Pods
that are not impacted, since we see some of those as well.  Not all
that clear to me exactly what triggers the exposure at the moment, but
as far as we can tell, shareProcessNamespace:true is required, and
then... something else.