blocked-edges/4.20.0-rc.3-PreRelease: Fixed in 4.20.0
To unblock 4.20.0's inclusion in the new candidate-4.22.

Also update hack/release-ga.sh to populate fixedIn for future
PreRelease risks.  Ideally only for the most recent prerelease.  But
while figuring out "ah, I'm on the last iteration of this while loop"
is possible in shell, the options I've found to it all seem pretty
fiddly.  So I'm just setting it for all the PreRelease risks.